RCAN Model for Task3a
Note : The model does not properly work right now since there are some issues with check pointing. It can be observed that after epoch49 there is a vanishing gradient problem (I think) which is why the output is blank. Since checkpointing is not working correctly, I could not extract the checkpoint with the best val_loss (Need to fix this)